Move CourseManagementService to Courses::ManagementService
Relocates the service to app/services/courses/management_service.rb
and namespaces it as Courses::ManagementService, consistent with other
services in that directory. Updates all references across controllers,
services, seeds, and specs.
